refactor: Entity 기반 레이어드 아키텍처 복원 및 Domain 호환성 유지
- ProductEntity: 비즈니스 로직(decreaseStock) 복원, Domain 변환 메서드 유지
- ProductService: Entity 기반으로 복원 (BigDecimal price, int stock)
- ProductController/Response/Request: Entity 기반으로 복원
- 기존 테스트 복원: ProductEntityTest, ProductRepositoryTest, ProductServiceTest, ProductPessimisticLockTest
- 새 테스트 추가: ProductTest (Domain 로직 단위 테스트)
- Testcontainers 설정: @DataJpaTest에 MySQL 컨테이너 설정 추가
- import 수정: ProductRepository → ProductJpaRepository

+++ server-java/src/main/java/kr/hhplus/be/server/presentation/product/ProductRequest.java
@@ -1,9 +1,11 @@
 package kr.hhplus.be.server.presentation.product;
 
+import java.math.BigDecimal;
+
 public record ProductRequest(
         String name,
         String description,
-        Long price,
-        Integer stock
+        BigDecimal price,
+        int stock
 ) {
 }
